Different Aggregation periods
Using `watson aggregate` is very handy when it comes to getting an
overview of tracked times over a given period. Unfortunately, this can
still be a bit overwhelming when aggregating larger periods of time.

This patch allows aggregation not only on a daily, but also on a weekly
or monthly basis. This makes it easier to review whats happening.